+++ axelor-base/src/main/java/com/axelor/apps/base/service/imports/listener/ImporterListener.java
@@ -21,7 +21,7 @@ import com.axelor.apps.base.exceptions.IExceptionMessage;
 import com.axelor.data.Listener;
 import com.axelor.db.Model;
 import com.axelor.exception.AxelorException;
-import com.axelor.exception.db.IException;
+import com.axelor.exception.db.repo.ExceptionOriginRepository;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
@@ -84,7 +84,7 @@ public class ImporterListener implements Listener {
             TraceBackRepository.TYPE_FUNCTIONNAL,
             I18n.get(IExceptionMessage.IMPORTER_LISTERNER_4),
             name),
-        IException.IMPORT);
+        ExceptionOriginRepository.IMPORT);
   }
 
   public boolean isImported() {