+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/batch/BatchReminderTimesheet.java
@@ -33,7 +33,7 @@ import com.axelor.apps.message.db.repo.MessageRepository;
 import com.axelor.apps.message.service.MessageService;
 import com.axelor.apps.message.service.TemplateMessageService;
 import com.axelor.auth.AuthUtils;
-import com.axelor.exception.db.IException;
+import com.axelor.exception.db.repo.ExceptionOriginRepository;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
@@ -100,7 +100,7 @@ public class BatchReminderTimesheet extends AbstractBatch {
         incrementDone();
       } catch (Exception e) {
         incrementAnomaly();
-        TraceBackService.trace(new Exception(e), IException.REMINDER, batch.getId());
+        TraceBackService.trace(new Exception(e), ExceptionOriginRepository.REMINDER, batch.getId());
       }
     }
   }
@@ -146,7 +146,8 @@ public class BatchReminderTimesheet extends AbstractBatch {
         incrementDone();
       } catch (Exception e) {
         incrementAnomaly();
-        TraceBackService.trace(new Exception(e), IException.INVOICE_ORIGIN, batch.getId());
+        TraceBackService.trace(
+            new Exception(e), ExceptionOriginRepository.INVOICE_ORIGIN, batch.getId());
       }
     }
   }
@@ -172,7 +173,7 @@ public class BatchReminderTimesheet extends AbstractBatch {
         incrementDone();
       } catch (Exception e) {
         incrementAnomaly();
-        TraceBackService.trace(new Exception(e), IException.REMINDER, batch.getId());
+        TraceBackService.trace(new Exception(e), ExceptionOriginRepository.REMINDER, batch.getId());
       }
     }
   }
@@ -209,7 +210,8 @@ public class BatchReminderTimesheet extends AbstractBatch {
         incrementDone();
       } catch (Exception e) {
         incrementAnomaly();
-        TraceBackService.trace(new Exception(e), IException.INVOICE_ORIGIN, batch.getId());
+        TraceBackService.trace(
+            new Exception(e), ExceptionOriginRepository.INVOICE_ORIGIN, batch.getId());
       }
     }
   }