+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/batch/BatchAccountingCutOff.java
@@ -26,7 +26,7 @@ import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.AccountingCutOffService;
 import com.axelor.db.JPA;
 import com.axelor.exception.AxelorException;
-import com.axelor.exception.db.IException;
+import com.axelor.exception.db.repo.ExceptionOriginRepository;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.google.inject.Inject;
@@ -109,7 +109,7 @@ public class BatchAccountingCutOff extends BatchStrategy {
           TraceBackService.trace(
               new AxelorException(
                   e, e.getCategory(), I18n.get("StockMove") + " %s", stockMove.getStockMoveSeq()),
-              IException.INVOICE_ORIGIN,
+              ExceptionOriginRepository.INVOICE_ORIGIN,
               batch.getId());
           incrementAnomaly();
           break;
@@ -117,7 +117,7 @@ public class BatchAccountingCutOff extends BatchStrategy {
           TraceBackService.trace(
               new Exception(
                   String.format(I18n.get("StockMove") + " %s", stockMove.getStockMoveSeq()), e),
-              IException.INVOICE_ORIGIN,
+              ExceptionOriginRepository.INVOICE_ORIGIN,
               batch.getId());
           incrementAnomaly();
           LOG.error("Anomaly generated for the stock move {}", stockMove.getStockMoveSeq());